1.9.10.8: Update Gradle wrapper, correct deprecated "version", add anim8 extension
The Gradle wrapper files were slightly out-of-date, mostly the JARs; the text files seem to differ only in newline placement. The "version" variable is deprecated for some reason in Gradle scripts, so the jar task now uses "archiveVersion". I hope anim8-gdx will be useful for making quick GIF previews of in-game visuals. Release 0.1.0 of anim8-gdx was just pushed to Maven Central, and should be available in a few hours.