Add a check to ensure there is a shared module
Without this check, an HTML module would also require a Shared module to build. With this check, the reference to the Shared module only appears when there are both an HTML and a Shared module. It looks like other modules don't need the source like GWT does, so this should be the only change needed.