Partial fix for HiDPI issues ( #17 )
The font isn't exactly how I want it, but I can now resize the window to much larger and at least that way increase the text size. The window chrome is back; no way to avoid it if users need to resize the window. https://i.imgur.com/ETDOODl.png