It's a MrStahlfelge party! And VisUI.
This has updates to MrStahlfelge's controllers repos, and keeps the now-split mapping and scene2d controllers deps separate from the now-legacy bigger gdx-controllerutils dep (which is now part of the official controllers extension, so controllerutils isn't needed). The czyzby websockets have been replaced with MrStahlfelge's websockets, which aren't 100% compatible but are very close. The small incompatibilities necessitated a change to the websocket template (3-5 lines in the websocket demo) , and testing that showed that visUI 1.4.6 was broken on GWT and needed to be updated to 1.4.8. I think that's enough for one commit...