Get rid of JCenter. Release 1.9.14.1 .
JCenter is closing down in just 3 months, so we should chuck it in the bin ASAP. This has already been done in libGDX itself. 1.9.14.1 updates both gdx-liftoff and generated projects to VisUI 1.4.10, which was released a few hours ago. There's also an odd typo fixed in the layout for the advanced tab, so it should look better.