Fix mistake in iOS resource code
This should help with #39 , and may resolve it entirely. The problem before was that instead of copying the file Contents.json from inside the path generator/ios/data/Media.xcassets/, it was copying that folder to the text of the JSON, which is the same as copying nothing. I've verified that Contents.json has text in it now, and it's the same as what it should be, I think, but I don't have a Mac to build on. I'll push a release shortly.