Release 1.9.14.2, defaulting to GDX 1.9.13
This release is mostly notable for iOS fixes, but it also defaults to libGDX 1.9.13 despite being in the 1.9.14.x release series. It's just... libGDX 1.9.14 shouldn't be used, it's too buggy. 1.9.13 works fine for now, and 1.9.15 should have all the current 1.9.14 problems worked out. gdx-liftoff also downgraded to 1.9.13 again, and this required downgrading vis-ui to 1.4.9, which still has the fixes we need.