Partially update to 1.10.1-SNAPSHOT while we are in a SNAPSHOT version too.
This brings LWJGL3 to 3.3.0 for Liftoff itself, but not for user projects yet. I encourage anyone who has an Apple Silicon Mac and uses libGDX to try to build Liftoff from source; it was painless here using `./gradlew jar` . If you ask nicely I can provide builds on the libGDX Discord, as well.