Bizarre fix needed for JitPack snapshot repos.
It looks like when a -SNAPSHOT version is fetched, it counts a build as being usable with a name that is the same as a normal commit hash, but with "-" as a prefix and "-1" (I hope that's all) as a suffix. This removes that prefix/suffix combination, if present, before using a version from JitPack.