Update gdx-teavm to 1.0.0-b2. Docs in generated.
I think it's a good idea to document (where people will see it) the TeaVM quirk where URL parameters break asset loading, since that just bit me testing this. The IDEA server is popular for GWT devs currently, and it just needs a tiny extra step to be usable for TeaVM too (removing the parameters).

Also, at least in TeaVM, the half-pixel position of the window in Scene2DTemplate messed up the initial appearance, so I rounded it.