Bring over Unicode username fix I found for...
SheerSt's pokewilds project, which (being reasonably popular, more-so than Liftoff) had encountered the Unicode username issue before. It actually turns out to be an issue for some ASCII punctuation as well. This only brings over template batch files, which need to have the correct version substituted in. That's another can of worms, since I want to do it with Gradle when I run the `jar` task...