Update TeaVM to 1.0.0-b8 .
This includes fixes for some small breaking changes in 1.0.0-b8 where packages changed. It also defaults to turning optimization on, since for projects that are just starting and are small, the time taken by optimization will be small. The comment about URL parameters doesn't seem to be needed any more, so it was removed.