Various important Android updates.
Target SDK is now always 34; sorry if that disrupts anyone making a new project, but Google wouldn't have allowed an SDK-33 project in anyway. The AGP version is now 8.1.4. The Android build.gradle file has (at least for now, and for Java) no warnings or errors. There are also small updates to the third-party libs digital and juniper.