Add gdx-UnBox2D, by Lyze.
It has to have a dependency on Box2D because GWT requires it to inherit Box2D (and ShapeDrawer). I figure Box2D is in the name, so most people probably want to use it with physics, but I don't know. I might be able to make the Box2D dependency conditional on a GWT project being present.