Catch and note exception if pref fails to save.
This shows the original Exception message as well because we don't have any other way to access the filename it is having trouble with.

This also notes in the README.md for the LWJGL3 module that it was once called 'desktop', in the hopes of clearing up some confusion there. This doesn't add any more text to the selection checkboxes, though.