Blindly trying to get MOE working.
I know this file was removed when MOE was, but it appears to still be needed. There are other files mentioned as missing that I haven't found yet. There is still more config that I can't even do because I don't have XCode, or a Mac. @Berstanio , your assistance would be appreciated!