Get Gradle tasks to run again!
There's currently no way to forget tasks this remembered from an earlier run unless you reset the preferences with the upper right button. Maybe a future version can have a "Reset Gradle Tasks" button when there are known Gradle tasks that could be reset. It's a problem for another day.