Weirdest. Fix. Ever.
This commit is mostly an edited revert of a few previous commits, but also updates to Construo 1.4.2. That is a nice version bump because it makes the plugin more flexible about how and where it can be applied, which turned out to be critical when Kotlin and Android are both in use. This also reverts the settings.gradle changes because they were causing some trouble and I wasn't sure how to solve it. Similar code may work here in the future.