Update to Gradle 8.13 !
This should fix the weird bug that superDev had in 8.12.1, I just need to undo the changes I made to HTML projects and hopefully things will work. This tries to make use of Daemon JVM auto-provisioning, but I can't really test that on a machine that has already been set up to use Gradle; it needs a totally fresh machine to see if it works correctly.