Use WinterAlexander's suggestion.
This should work for both Kotlin and Java templates, as well as Liftoff itself. There are lots of extra checks here to try to ensure no invalid chars get in at any point. The only possible issue would be two usernames having hashCode() collisions, and even that only matters if they're both logged on at the same time and both running a libGDX or LWJGL3 app with the same versions.