Attempt to address possible pref issues...
I think if the preferences store some invalid data, then they can reload that invalid data on later runs and cause problems with the validation not getting to check them. To try to alleviate that, I'm making the Java version default to 8 every time the app loads (which I know is a hassle if you use something other than 8 by default, but it's better than some crash or other issue happening when a typo'd version gets loaded, and reloaded). I also set the Gradle tasks to be empty because we can't currently remove them except by resetting the settings.

This also updates the version string to a SNAPSHOT.