Remove non-working action from #209 .
The action never completed successfully because softprops/action-gh-release@v2 was never permitted to be used in the libgdx organization. I don't have an especially easy time registering tags, and I think making releases via the browser interface makes the most sense to me anyway. So, we don't need the action, it looks like. Samples are still generated when a release is created.