Update Gradle, AGP, several 3rd-party deps.
Lombok, its Gradle plugin, gdx-gltf, and various "TEttinger libraries" have been updated. Fury has been updated with its new name, Fory, which is a backwards-incompatible change that is relatively easy to fix. There hasn't been an update to SquidWrath's stable releases in a while, so it might not work out-of-the-box with Fory's new name, package, class names, etc.