Switch to TextraTypist from Typing-Label.
We don't know when Typing-Label will next update, but when libGDX 1.14.0 comes out it will be incompatible due to using libGDX StringBuilder. It wouldn't be hard to update for that library, but if we want gdx-liftoff to update immediately we need to use something compatible (either scene2d.ui on its own, with no effects, or TextraTypist). It turned out TextraTypist was an easy swap-in!